chore: 기존 파일 삭제
