+++ src/@components/project/projectModal/ProjectModal.jsx
@@ -1,9 +1,10 @@
-import { Badge, Button } from '@goorm-dev/gds-components';
-import { GithubIcon } from '@goorm-dev/gds-icons';
-import React, { useEffect, useRef } from 'react';
 import * as S from './style';
+import React from 'react';
+import { Modal, ModalHeader, ModalBody, ModalFooter, Button } from '@goorm-dev/gds-components';
 
-export default function ProjectModal({
+function ProjectModal({
+  toggleModal,
+  isModalOpen,
   award,
   title,
   content,
@@ -15,99 +16,60 @@ export default function ProjectModal({
   design,
   frontend,
   backend,
-  setModalOpen,
 }) {
-  const modalRef = useRef();
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (modalRef.current && !modalRef.current.contains(event.target)) {
-        setModalOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [setModalOpen]);
-
+  console.log(frontend, backend);
   return (
-    <S.ModalContainer>
-      <S.Modal ref={modalRef}>
-        <S.ModalCloseBtn onClick={() => setModalOpen(false)}>X</S.ModalCloseBtn>
-        <S.ModalTitle>{title}</S.ModalTitle>
-        <S.ModalOverview>{content}</S.ModalOverview>
-        <S.BadgeWrapper>
-          <Badge color="dark" size="md" toggle={null}>
-            {award}
-          </Badge>
-          <Badge color="primary" size="md" toggle={null}>
-            {releaseLink === '' ? '' : '서비스 중'}
-          </Badge>
-        </S.BadgeWrapper>
-
-        <S.ModalPosterImg src={image} alt="project__poster-img" />
-        <S.ModalContent>
-          <S.ModalMemberWrapper>
-            <S.ModalMemberRowContent>
-              <S.ModalMemberRowContentDetail>
-                <S.PartText>기획 :&nbsp; </S.PartText>
-                <S.NameText> {pm}</S.NameText>
-              </S.ModalMemberRowContentDetail>
-              <S.ModalMemberRowContentDetail>
-                <S.PartText>디자인 :&nbsp; </S.PartText>
-                <S.NameText> {design}</S.NameText>
-              </S.ModalMemberRowContentDetail>
-            </S.ModalMemberRowContent>
-
-            <S.ModalMemberRowContent>
-              <S.ModalMemberRowContentDetail>
-                <S.PartText>프론트엔드 :&nbsp; </S.PartText>
-                <S.NameText>{frontend.join(', ')}</S.NameText>
-              </S.ModalMemberRowContentDetail>
-              <S.ModalMemberRowContentDetail>
-                <S.PartText>백엔드 :&nbsp; </S.PartText>
-                <S.NameText>{backend.join(', ')}</S.NameText>
-              </S.ModalMemberRowContentDetail>
-            </S.ModalMemberRowContent>
-          </S.ModalMemberWrapper>
-
-          <S.ButtonWrapper>
-            <Button
-              color="link"
-              iconSide="right"
-              size="xl"
-              tag="button"
-              onClick={() => window.open(backendLink, '_blank')}
-            >
-              <GithubIcon />
-              백엔드
+    <>
+      <Modal isOpen={isModalOpen} direction="vertical" size="md" toggle={toggleModal} type="center">
+        <ModalHeader toggle={toggleModal}></ModalHeader>
+        <ModalBody>
+          <S.ModalBodyContainer>
+            <S.ProjectContent>
+              <S.Title>{title}</S.Title>
+              <S.Description>{content}</S.Description>
+              <S.ProjectImg src={image} alt={title} />
+            </S.ProjectContent>
+            <S.ContributorContent>
+              <S.ContributorInfo>
+                <S.Role>기획</S.Role>
+                <S.Name className="paragraph-lg">{pm}</S.Name>
+              </S.ContributorInfo>
+              <S.ContributorInfo>
+                <S.Role>디자인</S.Role>
+                <S.Name className="paragraph-lg">{design}</S.Name>
+              </S.ContributorInfo>
+              <S.ContributorInfo className="developer">
+                <S.Role>프론트엔드</S.Role>
+                <S.Name className="paragraph-lg">{frontend.join(', ')}</S.Name>
+              </S.ContributorInfo>
+              <S.ContributorInfo className="developer">
+                <S.Role>백엔드</S.Role>
+                <S.Name className="paragraph-lg">{backend.join(', ')}</S.Name>
+              </S.ContributorInfo>
+            </S.ContributorContent>
+          </S.ModalBodyContainer>
+        </ModalBody>
+        <ModalFooter>
+          <React.Fragment key=".0">
+            <Button color="link" size="lg" onClick={() => window.open(backendLink, '_blank')}>
+              백엔드 Github
             </Button>
-            <Button
-              color="link"
-              iconSide="right"
-              size="xl"
-              tag="button"
-              onClick={() => window.open(frontendLink, '_blank')}
-            >
-              <GithubIcon />
-              프론트엔드
+            <Button color="link" size="lg" onClick={() => window.open(frontendLink, '_blank')}>
+              프론트엔드 Github
             </Button>
             <Button
               color="primary"
-              iconSide="right"
-              size="xl"
-              tag="button"
-              disabled={releaseLink === '' ? true : false}
+              size="lg"
+              disabled={releaseLink === ''}
               onClick={() => window.open(releaseLink, '_blank')}
             >
               서비스 바로가기
             </Button>
-          </S.ButtonWrapper>
-          {/* Team Members */}
-        </S.ModalContent>
-      </S.Modal>
-    </S.ModalContainer>
+          </React.Fragment>
+        </ModalFooter>
+      </Modal>
+    </>
   );
 }
+
+export default ProjectModal;
